+++ libipm/Makefile.am
@@ -19,6 +19,8 @@ libipm_la_SOURCES = \
   ercp.c \
   scp.h \
   scp.c \
+  scp_sync.h \
+  scp_sync.c \
   scp_application_types.h \
   scp_application_types.c
