+++ sesman/tools/sesrun.c
@@ -40,7 +40,8 @@
 #include "string_calls.h"
 #include "guid.h"
 
-#include "tools_common.h"
+#include "scp.h"
+#include "scp_sync.h"
 
 // cppcheck doesn't always set this macro to something in double-quotes
 #if defined(__cppcheck__)
@@ -453,7 +454,7 @@ handle_login_response(struct trans *t, int *server_closed)
 {
     enum scp_login_status login_result;
 
-    int rv = wait_for_sesman_reply(t, E_SCP_LOGIN_RESPONSE);
+    int rv = scp_sync_wait_specific(t, E_SCP_LOGIN_RESPONSE);
     if (rv != 0)
     {
         *server_closed = 1;
@@ -511,7 +512,7 @@ handle_create_session_response(struct trans *t)
     int display;
     struct guid guid;
 
-    int rv = wait_for_sesman_reply(t, E_SCP_CREATE_SESSION_RESPONSE);
+    int rv = scp_sync_wait_specific(t, E_SCP_CREATE_SESSION_RESPONSE);
     if (rv == 0)
     {
         rv = scp_get_create_session_response(t, &status,
